Adapt <ServerRouter> to use context.router
Added an `isServer` boolean to the history object provided by the <ServerRouter>. Replace will check `this.context.router.history.isServer` to determine whether or not it is a server render
